+++ HouseCrawler/AirflowAdmin/dags/NewHouseGuangzhou.py
@@ -152,7 +152,7 @@ for item in cur:
     buildingList_info_list.append(buildingList_info)
 
 random.shuffle(buildingList_info_list)
-index_skip = int(math.ceil(len(buildingList_info_list) / float(11))) + 1
+index_skip = int(math.ceil(len(buildingList_info_list) / float(20))) + 1
 for cur, index in enumerate(list(range(0, len(buildingList_info_list), index_skip))):
     t2 = PythonOperator(
         task_id='LoadBuildingListGuangzhou_%s' % cur,
